+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- Add parameter `skipSemVerEnsure` in app publish.
+
 ## [3.46.0] - 2019-08-26
 
 ## [3.46.0-beta.2] - 2019-08-26